+++ server/src/main/java/org/allaymc/server/world/service/AllayEntityPhysicsService.java
@@ -109,7 +109,6 @@ public class AllayEntityPhysicsService implements EntityPhysicsService {
                 return;
             }
 
-            // FIXME: fix bug caused by floating point number error here: 1.3f - 0.3f = 0.99999994
             var collidedBlocks = dimension.getCollidingBlockStates(entity.getOffsetAABB());
             if (collidedBlocks == null) {
                 // 1. The entity is not stuck in the block
@@ -605,7 +604,6 @@ public class AllayEntityPhysicsService implements EntityPhysicsService {
             }
         }
 
-        // FIXME: abs(coordinate) == Double.MAX_VALUE
         return coordinate;
     }